update accountHomePage.js - movie/tv show Header - responsivess
added a fontSize setting in the sx prop that changes depending on the screen size. For example, on smaller screens like xs, the font gets smaller.

So when the screen size is mobile (xs), the header text will show up at 1.5rem, and on bigger screens (md and up), it will be 3rem. You can change these values to whatever works best for you.